refactor: separate device invitation state from dialog views
