+++ src/components/device/AddDeviceDialog.tsx
@@ -1,27 +1,6 @@
-import {
-  AlertCircle,
-  Check,
-  CheckCircle2,
-  Clock,
-  Copy,
-  Info,
-  Loader2,
-  LockKeyhole,
-  RefreshCw,
-  XCircle,
-} from 'lucide-react'
-import { useEffect, useEffectEvent, useMemo, useRef, useState } from 'react'
+import { Check, Copy, Loader2, RefreshCw, XCircle } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import { getDeviceTrustSnapshot } from '@/api/daemon/device-trust'
-import {
-  cancelInvitation,
-  getSetupState,
-  issuePairingInvitation,
-  type CurrentInvitation,
-} from '@/api/daemon/setupV2'
-import { isUnlockSpaceError, unlockSpaceWithPassphrase } from '@/api/security'
-import type { SetupInvitationRevokedEvent } from '@/api/setupEvents'
-import { activeDeviceIds, findNewActiveDeviceId } from '@/components/device/pairing-success-utils'
+import { AddDeviceDialogBody } from '@/components/device/AddDeviceDialogBody'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
@@ -31,36 +10,13 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { Progress } from '@/components/ui/progress'
+import { useAddDeviceInvitation } from '@/hooks/useAddDeviceInvitation'
 import { useDialogSessionReset } from '@/hooks/useDialogSessionReset'
-import { daemonWs } from '@/lib/daemon-ws'
-import { formatInvitationCode } from '@/lib/invitation-code'
-import { createLogger } from '@/lib/logger'
-import { cn } from '@/lib/utils'
-
-const log = createLogger('add-device-dialog')
-
-// 默认邀请有效期 — 用于估算进度条百分比；倒计时仍按 expiresAtMs 显示真实剩余。
-const DEFAULT_TTL_MS = 5 * 60 * 1000
-// 配对成功后短暂展示成功态，再自动关闭对话框
-const SUCCESS_AUTO_CLOSE_MS = 5000
-
-type Step = 'credentials' | 'invitation' | 'success' | 'failed'
-type PairingCompletionTrigger = 'device_trust_changed' | 'refresh_required' | 'reconnected'
 
 interface AddDeviceDialogProps {
   open: boolean
   onOpenChange: (open: boolean) => void
   onSuccess?: () => void
-}
-
-function formatRemaining(ms: number): string {
-  const total = Math.max(0, Math.floor(ms / 1000))
-  const m = Math.floor(total / 60)
-  const s = total % 60
-  return `${m.toString().padStart(2, '0')}:${s.toString().padStart(2, '0')}`
 }
 
 function AddDeviceDialogInner({
@@ -70,420 +26,16 @@ function AddDeviceDialogInner({
   onOpenChangeComplete,
 }: AddDeviceDialogProps & { onOpenChangeComplete: (open: boolean) => void }) {
   const { t } = useTranslation()
-  const [invitation, setInvitation] = useState<CurrentInvitation | null>(null)
-  const [issuedAtMs, setIssuedAtMs] = useState<number | null>(null)
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [now, setNow] = useState(() => Date.now())
-  const [copied, setCopied] = useState(false)
-  const [step, setStep] = useState<Step>('invitation')
-  const [failureReason, setFailureReason] = useState<string | null>(null)
-  const [passphrase, setPassphrase] = useState('')
-  const initialDeviceIdsRef = useRef<ReadonlySet<string> | null>(null)
-  const pairingCompletionInFlightRef = useRef(false)
-  const reportSuccess = useEffectEvent(() => onSuccess?.())
-
-  // 倒计时 tick — 仅在邀请态 + 有邀请时启动
-  useEffect(() => {
-    if (!open || !invitation || step !== 'invitation') return
-    const id = setInterval(() => setNow(Date.now()), 1000)
-    return () => clearInterval(id)
-  }, [open, invitation, step])
-
-  // Restore the current invitation when opening, or issue one if necessary.
-  // Closing does not revoke it; local form state resets after the exit finishes.
-  // Keep translation updates out of the effect dependencies so resource reloads
-  // cannot issue another invitation or overwrite the completed state.
-  const reportIssueFailure = useEffectEvent(() => {
-    log.error({ error_kind: 'invitation_issue_failed' }, 'failed to load or issue invitation')
-    setError(t('devices.addDevice.errors.issueFailed'))
-  })
-  useEffect(() => {
-    if (!open) return
-    let cancelled = false
-    void (async () => {
-      setLoading(true)
-      setError(null)
-      try {
-        initialDeviceIdsRef.current = activeDeviceIds(await getDeviceTrustSnapshot())
-        const state = await getSetupState()
-        if (cancelled) return
-        if (state.currentInvitation) {
-          setInvitation(state.currentInvitation)
-          // 复用的邀请 — 没有真实"签发时间"，按 TTL 倒推一个估算值
-          setIssuedAtMs(state.currentInvitation.expiresAtMs - DEFAULT_TTL_MS)
-          log.info({ event: 'invitation_ready', mode: 'reused' }, 'pairing invitation ready')
-        } else if (state.rePairingRequired) {
-          setStep('credentials')
-          log.info({ event: 'credentials_required' }, 're-pairing credentials required')
-        } else {
-          const issued = await issuePairingInvitation()
-          if (cancelled) return
-          setInvitation(issued)
-          setIssuedAtMs(Date.now())
-          log.info({ event: 'invitation_ready', mode: 'standard' }, 'pairing invitation ready')
-        }
-      } catch {
-        if (cancelled) return
-        reportIssueFailure()
-      } finally {
-        if (!cancelled) setLoading(false)
-      }
-    })()
-    return () => {
-      cancelled = true
-    }
-  }, [open])
-
-  const confirmPairingCompleted = useEffectEvent(async (trigger: PairingCompletionTrigger) => {
-    if (
-      step !== 'invitation' ||
-      initialDeviceIdsRef.current === null ||
-      pairingCompletionInFlightRef.current
-    ) {
-      return
-    }
-    try {
-      const trust = await getDeviceTrustSnapshot()
-      const peerDeviceId = findNewActiveDeviceId(
-        initialDeviceIdsRef.current,
-        activeDeviceIds(trust)
-      )
-      if (peerDeviceId === null) return
-      pairingCompletionInFlightRef.current = true
-      try {
-        await cancelInvitation()
-      } catch {
-        log.warn(
-          { error_kind: 'invitation_cleanup_failed' },
-          'failed to clear completed invitation'
-        )
-      }
-      setStep('success')
-      log.info({ event: 'pairing_confirmed', trigger }, 'new device pairing confirmed')
-      reportSuccess()
-    } catch {
-      log.warn(
-        { error_kind: 'pairing_confirmation_failed', trigger },
-        'failed to verify completed invitation'
-      )
-      pairingCompletionInFlightRef.current = false
-    }
-  })
-
-  const handleInvitationRevoked = useEffectEvent((evt: SetupInvitationRevokedEvent) => {
-    if (step !== 'invitation' || loading) return
-    setFailureReason(evt.reason)
-    setStep('failed')
-    log.info({ event: 'invitation_revoked' }, 'pairing invitation revoked')
-  })
-
-  useEffect(() => {
-    if (!open) return
-    const unsubscribeEvents = daemonWs.subscribe(['device-trust', 'setup', 'system'], event => {
-      if (event.eventType === 'device-trust.changed') {
-        void confirmPairingCompleted('device_trust_changed')
-      } else if (event.eventType === 'system.refresh_required') {
-        void confirmPairingCompleted('refresh_required')
-      } else if (event.eventType === 'setup.invitationRevoked') {
-        handleInvitationRevoked(event.payload as SetupInvitationRevokedEvent)
-      }
-    })
-    const unsubscribeReconnect = daemonWs.onReconnect(
-      () => void confirmPairingCompleted('reconnected')
-    )
-
-    return () => {
-      unsubscribeEvents()
-      unsubscribeReconnect()
-    }
-  }, [open])
-
-  // 成功态自动关闭。把 onOpenChange 包成 useEffectEvent 移出依赖，避免父级
-  // 重渲染导致 setTimeout 被反复重建。
-  const closeDialog = useEffectEvent(() => onOpenChange(false))
-  useEffect(() => {
-    if (step !== 'success') return
-    const id = setTimeout(() => closeDialog(), SUCCESS_AUTO_CLOSE_MS)
-    return () => clearTimeout(id)
-  }, [step])
-
-  const remaining = invitation ? Math.max(0, invitation.expiresAtMs - now) : 0
-  const expired = invitation && step === 'invitation' ? remaining <= 0 : false
-  const totalMs = invitation && issuedAtMs ? invitation.expiresAtMs - issuedAtMs : DEFAULT_TTL_MS
-  const progress = invitation ? Math.max(0, Math.min(100, (remaining / totalMs) * 100)) : 0
-  const display = useMemo(
-    () => (invitation ? formatInvitationCode(invitation.code) : ''),
-    [invitation]
-  )
-
-  const handleCopy = async () => {
-    if (!invitation) return
-    try {
-      await navigator.clipboard.writeText(invitation.code)
-      setCopied(true)
-      setTimeout(() => setCopied(false), 1500)
-    } catch {
-      log.warn({ error_kind: 'clipboard_write_failed' }, 'clipboard.writeText failed')
-    }
-  }
-
-  const handleCancel = async () => {
-    setLoading(true)
-    try {
-      await cancelInvitation()
-    } catch {
-      log.warn({ error_kind: 'invitation_cancel_failed' }, 'cancelInvitation failed on close')
-    } finally {
-      setLoading(false)
-      onOpenChange(false)
-    }
-  }
-
-  const handleRegenerate = async () => {
-    setLoading(true)
-    setError(null)
-    setStep('invitation')
-    setFailureReason(null)
-    try {
-      initialDeviceIdsRef.current = activeDeviceIds(await getDeviceTrustSnapshot())
-      try {
-        await cancelInvitation()
-      } catch {
-        log.warn(
-          { error_kind: 'invitation_cancel_failed' },
-          'cancelInvitation before regenerate failed'
-        )
-      }
-      const issued = await issuePairingInvitation()
-      setInvitation(issued)
-      setIssuedAtMs(Date.now())
-      log.info({ event: 'invitation_ready', mode: 'regenerated' }, 'pairing invitation ready')
-    } catch {
-      log.error({ error_kind: 'invitation_issue_failed' }, 'regenerate invitation failed')
-      setError(t('devices.addDevice.errors.issueFailed'))
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const handleConfirmPassphrase = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    const value = passphrase.trim()
-    if (!value) return
-    setLoading(true)
-    setError(null)
-    log.info({ event: 'credentials_submitted' }, 're-pairing credentials submitted')
-    try {
-      await unlockSpaceWithPassphrase(value)
-      initialDeviceIdsRef.current = activeDeviceIds(await getDeviceTrustSnapshot())
-      const issued = await issuePairingInvitation()
-      setInvitation(issued)
-      setIssuedAtMs(Date.now())
-      setPassphrase('')
-      setStep('invitation')
-      log.info(
-        { event: 'invitation_ready', mode: 'legacy_re_pairing' },
-        're-pairing invitation ready'
-      )
-    } catch (err) {
-      const wrongPassphrase = isUnlockSpaceError(err) && err.code === 'WRONG_PASSPHRASE'
-      if (wrongPassphrase) {
-        log.info(
-          { error_kind: 'wrong_passphrase', event: 'credentials_rejected' },
-          're-pairing credentials rejected'
-        )
-      } else {
-        log.warn(
-          { error_kind: 'credentials_confirmation_failed', event: 'credentials_rejected' },
-          're-pairing credentials rejected'
-        )
-      }
-      setError(
-        wrongPassphrase
-          ? t('devices.addDevice.rePairing.wrongPassphrase')
-          : t('devices.addDevice.rePairing.failed')
-      )
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const failureMessage = useMemo(() => {
-    if (!failureReason) return t('devices.addDevice.failed.unknown')
-    const key = `devices.addDevice.failed.reasons.${failureReason}`
-    const translated = t(key)
-    if (translated !== key) return translated
-    return t('devices.addDevice.failed.fallback', { reason: failureReason })
-  }, [failureReason, t])
-
-  // ── 主体内容 ─────────────────────────────────────────
-  let body: React.ReactNode = null
-  if (step === 'credentials') {
-    body = (
-      <form
-        data-testid="re-pairing-passphrase-step"
-        className="flex flex-col gap-4 py-3"
-        onSubmit={handleConfirmPassphrase}
-      >
-        <div className="flex items-start gap-3 rounded-lg border border-primary/20 bg-primary/5 p-3 text-sm text-foreground">
-          <LockKeyhole className="mt-0.5 size-4 shrink-0 text-primary" />
-          <span>{t('devices.addDevice.rePairing.description')}</span>
-        </div>
-        <div className="flex flex-col gap-2">
-          <Label htmlFor="re-pairing-passphrase">
-            {t('devices.addDevice.rePairing.passphraseLabel')}
-          </Label>
-          <Input
-            id="re-pairing-passphrase"
-            type="password"
-            autoComplete="current-password"
-            autoFocus
-            value={passphrase}
-            onChange={event => setPassphrase(event.target.value)}
-            placeholder={t('devices.addDevice.rePairing.passphrasePlaceholder')}
-            disabled={loading}
-          />
-        </div>
-        {error && <p className="text-sm font-medium text-destructive">{error}</p>}
-        <Button
-          type="submit"
-          data-testid="re-pairing-confirm-passphrase"
-          disabled={loading || !passphrase.trim()}
-        >
-          {loading && <Loader2 className="mr-2 size-4 animate-spin" />}
-          {loading
-            ? t('devices.addDevice.rePairing.submitting')
-            : t('devices.addDevice.rePairing.submit')}
-        </Button>
-      </form>
-    )
-  } else if (step === 'success') {
-    body = (
-      <div data-testid="add-device-success" className="flex flex-col items-center gap-3 py-8">
-        <div className="flex size-14 items-center justify-center rounded-full bg-emerald-500/15 text-emerald-600 dark:text-emerald-400">
-          <CheckCircle2 className="size-8" />
-        </div>
-        <div className="text-center">
-          <p className="text-base font-semibold text-foreground">
-            {t('devices.addDevice.success.title')}
-          </p>
-          <p className="mt-1 text-sm text-muted-foreground">
-            {t('devices.addDevice.success.subtitle')}
-          </p>
-        </div>
-      </div>
-    )
-  } else if (step === 'failed') {
-    body = (
-      <div className="flex flex-col items-center gap-3 py-6">
-        <div className="flex size-12 items-center justify-center rounded-full bg-destructive/15 text-destructive">
-          <AlertCircle className="size-7" />
-        </div>
-        <div className="text-center">
-          <p className="text-base font-semibold text-foreground">
-            {t('devices.addDevice.failed.title')}
-          </p>
-          <p className="mt-1 text-sm text-muted-foreground">{failureMessage}</p>
-          <p className="mt-3 text-xs text-muted-foreground/70">
-            {t('devices.addDevice.failed.networkHint')}
-          </p>
-        </div>
-      </div>
-    )
-  } else if (loading && !invitation) {
-    body = (
-      <div className="flex items-center justify-center gap-3 py-12 text-sm text-muted-foreground">
-        <Loader2 className="size-4 animate-spin" />
-        {t('devices.addDevice.loading')}
-      </div>
-    )
-  } else if (error && !invitation) {
-    body = (
-      <div className="flex flex-col items-center gap-3 py-10">
-        <p className="text-sm text-destructive">{error}</p>
-        <Button variant="outline" size="sm" onClick={handleRegenerate} disabled={loading}>
-          <RefreshCw className="mr-2 size-3.5" />
-          {t('devices.addDevice.actions.regenerate')}
-        </Button>
-      </div>
-    )
-  } else if (invitation) {
-    body = (
-      <div className="space-y-4 py-2">
-        {/* 邀请码主卡 — 视觉焦点 */}
-        <div
-          className={cn(
-            'rounded-2xl border bg-gradient-to-br p-5 transition-colors',
-            expired
-              ? 'border-destructive/30 from-destructive/5 to-transparent'
-              : 'border-primary/20 from-primary/[0.04] to-transparent'
-          )}
-        >
-          <div
-            data-testid="add-device-invitation-code"
-            className={cn(
-              'select-all text-center font-mono font-semibold tabular-nums text-foreground',
-              'text-[28px] tracking-[0.18em] sm:text-[32px] sm:tracking-[0.2em]',
-              expired && 'text-muted-foreground/50 line-through decoration-1'
-            )}
-            aria-label={invitation.code}
-          >
-            {display}
-          </div>
-
-          <div className="mt-5 space-y-2">
-            <Progress
-              value={progress}
-              className={cn('h-1', expired && '[&>[data-slot=progress-indicator]]:bg-destructive')}
-            />
-            <div
-              className={cn(
-                'flex items-center justify-center gap-1.5 text-xs tabular-nums',
-                expired ? 'text-destructive' : 'text-muted-foreground'
-              )}
-            >
-              <Clock className="size-3" />
-              {expired
-                ? t('devices.addDevice.expired')
-                : t('devices.addDevice.expiresIn', {
-                    remaining: formatRemaining(remaining),
-                  })}
-            </div>
-          </div>
-        </div>
-
-        {/* 提示：还需空间口令 */}
-        <div className="flex items-start gap-2.5 rounded-lg bg-muted/50 px-3.5 py-2.5 text-xs text-muted-foreground">
-          <Info className="mt-0.5 size-3.5 shrink-0" />
-          <span className="leading-relaxed">{t('devices.addDevice.passphraseHint')}</span>
-        </div>
-      </div>
-    )
-  }
+  const invitationState = useAddDeviceInvitation({ open, onOpenChange, onSuccess })
+  const { invitation, loading, step, copied, expired, handleCopy, handleCancel, handleRegenerate } =
+    invitationState
 
   // ── Footer ───────────────────────────────────────────
   let footer: React.ReactNode
   if (step === 'credentials' || step === 'success') {
     // 自动关闭，无按钮
     footer = null
-  } else if (step === 'failed') {
-    footer = (
-      <>
-        <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>
-          {t('devices.addDevice.actions.close')}
-        </Button>
-        <Button onClick={handleRegenerate} disabled={loading}>
-          {loading ? (
-            <Loader2 className="mr-2 size-4 animate-spin" />
-          ) : (
-            <RefreshCw className="mr-2 size-4" />
-          )}
-          {t('devices.addDevice.actions.regenerate')}
-        </Button>
-      </>
-    )
-  } else if (expired) {
+  } else if (step === 'failed' || expired) {
     footer = (
       <>
         <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>
@@ -558,7 +110,7 @@ function AddDeviceDialogInner({
           )}
         </DialogHeader>
 
-        {body}
+        <AddDeviceDialogBody invitationState={invitationState} />
 
         {footer && <DialogFooter>{footer}</DialogFooter>}
       </DialogContent>
